Improved collision, scoring, lives, HUD, data-driven levels, tests, CI, and code quality
- Fixed ball-brick collision to properly detect which face was hit
- Fixed paddle collision to clamp ball above paddle
- Added score tracking (100pts x level per brick) and 3-life system
- Added HUD showing Level, Score, and Lives
- Added signal handlers for SIGTERM, SIGQUIT, SIGSEGV
- Upgraded ConsoleColor to enum class for type safety
- Removed using namespace std::chrono; from Main.cpp
- Levels now load from levels/levelN.txt files with hardcoded fallback
- Added make release target with -O2 and strip
- Added unit tests (make test): Ball wall bouncing, Box::Contains
- Added GitHub Actions CI workflow
- Fixed level 3 bottom row extending past border

+++ BaseObject.cpp
@@ -4,6 +4,6 @@
 void BaseObject::Draw() const
 {
 	Console::SetCursorPosition(x_position, y_position);
-	Console::ForegroundColor(color);
+	Console::ForegroundColor(static_cast<WORD>(color));
 	std::cout << visage;
 }
